autogen operants base
