+++ paddle/phi/api/yaml/generator/tensor_gen.py
@@ -30,6 +30,35 @@ inplace_optional_out_type_map = {
 indent = "  "
 
 
+operants_base_include = """// Generated by paddle/phi/api/yaml/generator/tensor_gen.py
+
+#pragma once
+
+#include "paddle/phi/api/include/tensor.h"
+
+"""
+
+operants_base_start = """
+namespace paddle {
+
+namespace operants {
+
+using Tensor = paddle::experimental::Tensor;
+
+class TensorOperantsBase {
+ public:
+  virtual ~TensorOperantsBase() = default;
+"""
+
+
+operants_base_end = """};
+
+}  // namespace operants
+}  // namespace paddle
+
+"""
+
+
 operants_header_include = """// Generated by paddle/phi/api/yaml/generator/tensor_gen.py
 
 #pragma once
@@ -94,6 +123,17 @@ class OperantsAPI(ForwardAPI):
         if self.get_api_func_name() in prims:
             self.is_prim_api = True
 
+    def gene_operants_base(self):
+        api_func_name = self.get_api_func_name()
+        if api_func_name[-1] != '_':
+            return f"""
+{indent}virtual {self.get_return_type()} {api_func_name}({self.get_declare_args()}) = 0;
+"""
+        else:
+            return f"""
+{indent}virtual {self.get_return_type(inplace_flag=True)} {api_func_name}({self.get_declare_args(inplace_flag=True)}) = 0;
+"""
+
     def gene_operants_declaration(self):
         api_func_name = self.get_api_func_name()
         if api_func_name[-1] != '_':
@@ -141,12 +181,12 @@ def generate_tensor_operants_api(
             if api_list:
                 apis.extend(api_list)
 
-    # operants_base_file = open(operants_base_path, 'w')
+    operants_base_file = open(operants_base_path, 'w')
     operants_header_file = open(operants_header_path, 'w')
     operants_source_file = open(operants_source_path, 'w')
 
-    # operants_base_file.write(operants_base_include)
-    # operants_base_file.write(operants_base_start)
+    operants_base_file.write(operants_base_include)
+    operants_base_file.write(operants_base_start)
     operants_header_file.write(operants_header_include)
     operants_header_file.write(operants_header_start)
     operants_source_file.write(operants_source_include)
@@ -160,14 +200,17 @@ def generate_tensor_operants_api(
     for api in apis:
         operants_api = OperantsAPI(api, api_prims)
         if operants_api.is_prim_api:
+            operants_base_file.write(operants_api.gene_operants_base())
             operants_header_file.write(operants_api.gene_operants_declaration())
             operants_source_file.write(
                 operants_api.gene_operants_implementation()
             )
 
+    operants_base_file.write(operants_base_end)
     operants_header_file.write(operants_header_end)
     operants_source_file.write(operants_source_end)
 
+    operants_base_file.close()
     operants_header_file.close()
     operants_source_file.close()
 
@@ -210,10 +253,10 @@ def main():
     options = parser.parse_args()
 
     api_yaml_path = options.api_yaml_path
-    api_prim_yaml_path = options.api_prim_yaml_path
     operants_base_path = options.operants_base_path
     operants_header_path = options.phi_tensor_operants_header_path
     operants_source_path = options.phi_tensor_operants_source_path
+    api_prim_yaml_path = options.api_prim_yaml_path
 
     generate_tensor_operants_api(
         api_yaml_path,
